--- apps/wsb-task/src/commands/action/mod.rs
@@ -1,2 +1,0 @@
-mod output;
-pub use output::ActionOutput;
