+++ apps/wsb-task/src/commands/cargo/build_pilot/params.rs
@@ -1,6 +1,6 @@
 use crate::commands::support::CCFindable;
-use crate::commands::ActionOutput;
 use crate::core::targets::{AsBuildTarget, BuildTarget};
+use crate::core::ActionOutput;
 use std::path::PathBuf;
 
 #[derive(Debug)]
