+++ apps/wsb-task/src/commands/support/definable.rs
@@ -1,4 +1,4 @@
-use crate::commands::Action;
+use crate::core::Action;
 use crate::TaskResult;
 use shellwork::core::command;
 use shellwork::core::command::{Runner, Unprepared};
