+++ apps/wsb-task/src/core/mod.rs
@@ -1,2 +1,6 @@
+mod action;
+pub use action::Action;
+pub use action::ActionOutput;
+
 #[macro_use]
 pub mod targets;
